FreezyEvaluator

Description

FreezyEvaluator - The Evaluator for Freezy Expressions

Evaluator Types

The Environment

globalEnv :: Env

Constructor for the global environment

openScope :: Env -> Env

Opens a new environment inside of env

assign :: String -> FreezyValue -> Evaluator FreezyValue

Assignment. This doesn't allow mutation.

lookup :: String -> Evaluator FreezyValue

Lookup a variable in the environment. This does

first check the current scope and return a result if found

get the enclosing scope and repeat with step one or

error out if the variable cannot be found and we reached the root scope

Evaluator Stack

type Evaluator a = StateT Env (ExceptT EvaluatorError IO) a

Evaluator Monad

runEvaluator :: Env -> Evaluator a -> IO (Either EvaluatorError (a, Env))

Unwrapping of the monad stack

evalEvaluator :: Env -> Evaluator a -> IO (Either EvaluatorError a)

Unwrap the monad stack while discarding the state

data EvaluatorError

Type for Evaluator errors

Constructors

EvaluatorError

Fields

reason :: String

e_line :: Int

Evalution Logic

evaluateIt :: Program -> Evaluator ()

Evalauates a FreezyProgram

evaluate :: Expr -> Evaluator FreezyValue

Evalauates a single Freezy Expression

evaluateBlock :: Maybe FreezyValue -> [Expr] -> Evaluator FreezyValue

Helper function to evaluate a function body/block expr

insertArgs :: [Token] -> [FreezyValue] -> Evaluator ()

Helper function to put the arguments in the environment

Operators

performUnaryOp :: Token -> FreezyValue -> Evaluator FreezyValue

perform a unary operation and do some rudimentary type checking

performOp :: Token -> FreezyValue -> FreezyValue -> Evaluator FreezyValue

dispatch a binary operation to either the numericOp handler or the polyOp handler

numericOp :: TokenType -> FreezyValue -> FreezyValue -> Evaluator FreezyValue

check the operand types and dispatch to one of the numerical binary operations

polyOp :: TokenType -> FreezyValue -> FreezyValue -> Evaluator FreezyValue

dispatch to one of the polymorphic binary operations

Helper Functions

litToVal :: Token -> Evaluator FreezyValue

Converts a literal token to its runtime representation