FreezyLang

Description

A Freezy program might look like this:

   fun plus(a, b,) {
       a + b
   }

   # this is a comment :)

   let a = fn (a,b){a+b}(1,4)
   print a # prints 5

   if plus( 2, 5) > 7
       then "wut"
       else "jay"

   let a = 10
   let name = "peter"

   print name
   print a;
   print name ~ a
   

Semicolon might be ommited Dangling Commas are allowed

Documentation

data TokenType

Freezy syntax tokens

Constructors

IDENTIFIER
Identifier

NUMBER
Literals

STRING

TRUE
Booleans

FALSE

IF
Conditional branching

THEN

ELSE

LET
Assignment Function declarations Lambdas

EQUAL

FUN

FN

PRINT
Print

PLUS
Math operators

MINUS

DASH

STAR

GR_EQ
Comparison operatos

LE_EQ

GR

LE

EQ_EQ
Equality check operators

BANG_EQ

TILDE
String Concatenation Operator

BANG
Not operator

COMMA
Misc

SEMICOLON

LPAR

RPAR

LBRACE

RBRACE

EOF
Special EOF token

WS
Throwaway token for whitespace/comments

data Token

The full token type with additional information

Constructors

Token

Fields

t_type :: TokenType

t_lexeme :: String

The lexeme for literals

t_line :: Int

the line of the token is tracked for error reporting

data Expr

Freezy Expressions

Constructors

Unary Token Expr
Unary minus and not

Binary Expr Token Expr
Binary Operations

Literal Token
String and Number Literals

IfExpr Expr Expr Expr
Conditional. Both then and else are mandatory

Call Expr [Expr]
Function Call

Fn [Token] [Expr]
Lambdas

Let Token Expr
Constant definition

Const Token
Constant use

Grouping Expr
Groupings

Print Expr
evaluates to the printed string

type Program = [Expr]

A FreezyProgram is just a list of Expressions. This is

the output of the main parsing routine

the input to the main interpreter

type Env = Map String FreezyValue

a simple Environment for the Evaluator.

this goes here because we store a closure (Env) in Function Values.

data FreezyValue

Freezy Runtime Values

Constructors

Number Integer
Unbounded Integers

String String
Strings

Boolean Bool
Boolean

Function Env [Token] [Expr]
Functions