FreezyOperators

Description

Implementation of the operators for Freezy

Numeric Operators. The Typechecking is done in the Evaluator

Unary

unaryMinus :: FreezyValue -> FreezyValue

unary Minus is the only unary numeric op in Freezy

Binary Math

plus :: FreezyValue -> FreezyValue -> FreezyValue

adds to numbers

minus :: FreezyValue -> FreezyValue -> FreezyValue

subtract one number from another

times :: FreezyValue -> FreezyValue -> FreezyValue

multiply two numbers

division :: FreezyValue -> FreezyValue -> FreezyValue

perform integer division

Binary Comparison

greater :: FreezyValue -> FreezyValue -> FreezyValue

Check wether a number is greater than another

only numbers can be compared in Freezy

greaterEqual :: FreezyValue -> FreezyValue -> FreezyValue

Check wether a number is greater or equal than another

only numbers can be compared in Freezy

lesser :: FreezyValue -> FreezyValue -> FreezyValue

Check wether a number is lesser than another

only numbers can be compared in Freezy

lesserEqual :: FreezyValue -> FreezyValue -> FreezyValue

Check wether a number is lesser or equal than another

only numbers can be compared in Freezy

Polymorphic Operators

unaryNot :: FreezyValue -> FreezyValue

Falsify a value. See isTruthy for a definition of what is considered to be true

equal :: FreezyValue -> FreezyValue -> FreezyValue

Equality Checks Functions and Lambdas are not comparable everything else is compared by value

notEqual :: FreezyValue -> FreezyValue -> FreezyValue

Check wether two values are unequal

concatenate :: FreezyValue -> FreezyValue -> FreezyValue

(Stringify) and concatenate two values

Helper

stringify :: FreezyValue -> String

Turn a value into a string. Functions and Lambdas are not really stringifiable

isTruthy :: FreezyValue -> Bool

Checks wether a given FreezyValue is truthy. In Freezy everything besides '0' and false is considered to be truthy