FreezyParser

Description

FreezyParser - A Parser for FreezyLang

Parser Types

type Parser a = ExceptT ParserError (StateT ParserState Identity) a

Parser Monad

runParser :: ParserState -> Parser a -> (Either ParserError a, ParserState)

Unwrapping of the monad stack

data ParserState

State of the parser

Constructors

ParserState

Fields

tokens :: [Token]

previousTokens :: [Token]

initParserState :: [Token] -> ParserState

Initiate the parse state with a list of tokens

data ParserError

Type for parser errors

Constructors

ParserError

Fields

reason :: String

e_line :: Int

Main Parser Logic

expression :: Parser Expr

Entry point to the recursive descent expression parser

concatenation :: Parser Expr

equality :: Parser Expr

comparison :: Parser Expr

term :: Parser Expr

factor :: Parser Expr

unary :: Parser Expr

ifExpr :: Parser Expr

assignment :: Parser Expr

printExpr :: Parser Expr

primary :: Parser Expr

binaryParser :: Parser Expr -> [TokenType] -> Parser Expr

generate a Binary parser for the given operators

Parsing Primitives

isAtEnd :: Parser Bool

check if we are at the end

advance :: Parser Token

advance the state by one token

match :: [TokenType] -> Parser Bool

advance if the current token matches any of the given TokenTypes

check :: TokenType -> Parser Bool

check if the current token matches the given TokenType

consume :: TokenType -> String -> Parser Token

consume the current token if it matches the given TokenType or throw an error

peek :: Parser Token

returns the current token

previous :: Parser Token

returns the previous token